reorganize some classes
