+++ app/src/main/java/com/example/taylor/gifbox/activity/MainActivity.kt
@@ -1,10 +1,12 @@
-package com.example.taylor.gifbox
+package com.example.taylor.gifbox.activity
 
 import android.arch.lifecycle.Observer
 import android.arch.lifecycle.ViewModelProviders
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.support.v7.widget.LinearLayoutManager
+import com.example.taylor.gifbox.R
+import com.example.taylor.gifbox.adapter.GifAdapter
 import com.example.taylor.gifbox.viewmodel.MainActivityVM
 import kotlinx.android.synthetic.main.activity_main.*
