+++ app/src/main/java/com/example/taylor/gifbox/controller/GiphyApi.kt
@@ -1,4 +1,4 @@
-package com.example.taylor.gifbox
+package com.example.taylor.gifbox.controller
 
 import com.example.taylor.gifbox.response.GifListResponse
 import io.reactivex.Observable
@@ -8,11 +8,10 @@ import retrofit2.http.Query
 /**
  * Created by Taylor on 1/28/2018.
  */
-const val TRENDING_URL = "gifs/trending"
 
 interface GiphyApi {
 
-    @GET(TRENDING_URL)
+    @GET("gifs/trending")
     fun fetchTrending(@Query("api_key") apiKey: String,
                       @Query("limit") limit : Int = 10,
                       @Query("rating") rating: String = "R",
