+++ app/src/main/java/com/example/taylor/gifbox/module/ApiModule.kt
@@ -1,9 +1,8 @@
 package com.example.taylor.gifbox.module
 
-import com.example.taylor.gifbox.GiphyApi
+import com.example.taylor.gifbox.controller.GiphyApi
 import com.example.taylor.gifbox.controller.ApiController
 import com.example.taylor.gifbox.controller.ApiControllerImpl
-import com.example.taylor.gifbox.model.PaginationObject
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import dagger.Module
