+++ app/src/main/java/com/example/taylor/gifbox/module/DataModule.kt
@@ -1,7 +1,6 @@
 package com.example.taylor.gifbox.module
 
 import android.app.Application
-import android.arch.persistence.db.SupportSQLiteDatabase
 import android.arch.persistence.room.Database
 import android.arch.persistence.room.Room
 import android.arch.persistence.room.RoomDatabase
