+++ app/src/test/java/com/example/taylor/gifbox/viewmodel/MainActivityVMTest.kt
@@ -11,7 +11,7 @@ import com.nhaarman.mockitokotlin2.whenever
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
 import org.junit.Before
-import org.junit.Rule
+import org.junit.ClassRule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mock
@@ -21,7 +21,7 @@ import org.mockito.junit.MockitoJUnitRunner
 class MainActivityVMTest {
 
     companion object {
-        @Rule
+        @ClassRule
         @JvmField
         val rule = InstantTaskExecutorRule()
     }
